solutions/aisearch-document-authentication/requirements.txt
python-dotenv
openai 
azure-search-documents
azure-identity
azure-storage-blob
pdfplumber
tiktoken